% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{bpi_price}
\alias{bpi_price}
\title{Query the current Bitcoin Price Index.}
\usage{
bpi_price(currency = NULL)
}
\arguments{
\item{currency}{A list of currency code (default is USD).}
}
\value{
Return a dataframe consists of current BPI.
}
\description{
This function is used to query the current price of BPI.
Currency can be specified by adding a list of currency codes.
The return dataframe consists of four columns data including currency code, description, float rate and time.
Powered by CoinDesk: https://www.coindesk.com/price/
}
\examples{
bpi_price()
bpi_price(c("EUR","CNY"))
}
