% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{get_currency_code}
\alias{get_currency_code}
\title{Get the currency code for Bitcoin Price Index query.}
\usage{
get_currency_code()
}
\value{
Return a dataframe consists of supported currency codes.
}
\description{
This function is used to show all the currency codes used for BPI query.
Powered by CoinDesk: https://www.coindesk.com/price/
}
\examples{
get_currency_code()
}
