% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{get_history}
\alias{get_history}
\title{Query the history Bitcoin Price Index.}
\usage{
get_history(currency = NULL, start = NULL, end = NULL,
  yesterday = FALSE)
}
\arguments{
\item{currency}{Query currency code (default is USD).}

\item{start}{Query start date, YYYY-MM-DD format.(default is last one month)}

\item{end}{Query end date, YYYY-MM-DD format.}

\item{yesterday}{Logical variable, if set to true, will return yesterday's BPI.}
}
\value{
Return a dataframe consists of historical BPI.
}
\description{
This function is used to query the history BPI.
Currency, start date, end date can be specified in the query.
The return dataframe consists of BPI data and respective date.
Powered by CoinDesk: https://www.coindesk.com/price/
}
\examples{
get_history('CNY','2017-10-11','2017-11-20')
get_history(yesterday=TRUE)
get_history('CNY')
get_history()
}
