% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bpi}
\alias{plot_bpi}
\title{Plot the historical data of BPI.}
\usage{
plot_bpi(df)
}
\arguments{
\item{df}{Query result from history BPI}
}
\description{
This function is used to draw a simple plot for historical BPI.
Powered by CoinDesk: https://www.coindesk.com/price/
}
\examples{
plot_bpi(df)
}
